+++ zebra-rpc/src/methods/types/get_block_template.rs
@@ -3,6 +3,7 @@
 pub mod constants;
 pub mod parameters;
 pub mod proposal;
+pub mod tachyon;
 pub mod zip317;
 
 #[cfg(test)]